Merge 3.5.9

+++ src/mongo/s/write_ops/batch_write_op.cpp
@@ -405,13 +405,16 @@ void BatchWriteOp::buildBatchRequest(const TargetedWriteBatch& targetedBatch,
     request->setNS(_clientRequest.getNS());
     request->setShouldBypassValidation(_clientRequest.shouldBypassValidation());
 
-    const vector<TargetedWrite*>& targetedWrites = targetedBatch.getWrites();
+    const auto batchType = _clientRequest.getBatchType();
+    const auto batchTxnNum = _clientRequest.getTxnNum();
 
-    for (vector<TargetedWrite*>::const_iterator it = targetedWrites.begin();
-         it != targetedWrites.end();
-         ++it) {
-        const WriteOpRef& writeOpRef = (*it)->writeOpRef;
-        BatchedCommandRequest::BatchType batchType = _clientRequest.getBatchType();
+    boost::optional<std::vector<int32_t>> stmtIdsForOp;
+    if (batchTxnNum) {
+        stmtIdsForOp.emplace();
+    }
+
+    for (auto& targetedWrite : targetedBatch.getWrites()) {
+        const WriteOpRef& writeOpRef = targetedWrite->writeOpRef;
 
         // NOTE:  We copy the batch items themselves here from the client request
         // TODO: This could be inefficient, maybe we want to just reference in the future
@@ -424,19 +427,24 @@ void BatchWriteOp::buildBatchRequest(const TargetedWriteBatch& targetedBatch,
             BatchedUpdateDocument* updateDoc = new BatchedUpdateDocument;
             clientUpdateRequest->getUpdatesAt(writeOpRef.first)->cloneTo(updateDoc);
             request->getUpdateRequest()->addToUpdates(updateDoc);
-        } else {
-            dassert(batchType == BatchedCommandRequest::BatchType_Delete);
+        } else if (batchType == BatchedCommandRequest::BatchType_Delete) {
             BatchedDeleteRequest* clientDeleteRequest = _clientRequest.getDeleteRequest();
             BatchedDeleteDocument* deleteDoc = new BatchedDeleteDocument;
             clientDeleteRequest->getDeletesAt(writeOpRef.first)->cloneTo(deleteDoc);
             request->getDeleteRequest()->addToDeletes(deleteDoc);
+        } else {
+            MONGO_UNREACHABLE;
+        }
+
+        if (stmtIdsForOp) {
+            stmtIdsForOp->push_back(_clientRequest.getStmtIdForWriteAt(writeOpRef.first));
         }
 
         // TODO: We can add logic here to allow aborting individual ops
         // if ( NULL == response ) {
         //    ->responses.erase( it++ );
         //    continue;
-        //}
+        // }
     }
 
     if (_clientRequest.isWriteConcernSet()) {
@@ -454,6 +462,11 @@ void BatchWriteOp::buildBatchRequest(const TargetedWriteBatch& targetedBatch,
     }
 
     request->setShardVersion(targetedBatch.getEndpoint().shardVersion);
+
+    if (batchTxnNum) {
+        request->setTxnNum(batchTxnNum);
+        request->setStmtIds(std::move(stmtIdsForOp));
+    }
 }
 
 void BatchWriteOp::noteBatchResponse(const TargetedWriteBatch& targetedBatch,
